Refactor : SseService Refactoring
