--- src/main/java/com/sdy/bbbb/SSE/SseService2.java
@@ -1,92 +1,0 @@
-package com.sdy.bbbb.SSE;
-
-import com.sdy.bbbb.entity.Account;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
-
-import java.io.IOException;
-import java.util.List;
-import java.util.Map;
-
-@Service
-@RequiredArgsConstructor
-public class SseService2 {
-
-    private final EmitterRepository emitterRepository = new EmitterRepositoryImpl();
-    private final NotificationRepository notificationRepository;
-    private static final Long DEFAULT_TIMEOUT = 60 * 1000L;
-
-    public SseEmitter subscribe(Long memberId) {
-//        lastEventId = "";
-        String emitterId = makeTimeIncludeId(memberId);
-        SseEmitter emitter = emitterRepository.save(emitterId, new SseEmitter(DEFAULT_TIMEOUT));
-        emitter.onCompletion(() -> emitterRepository.deleteById(emitterId));
-        emitter.onTimeout(() -> emitterRepository.deleteById(emitterId));
-
-        // 503 에러를 방지하기 위한 더미 이벤트 전송
-        String eventId = makeTimeIncludeId(memberId);
-        sendNotification(emitter, eventId, emitterId, "EventStream Created. [userId=" + memberId + "]");
-
-        // 클라이언트가 미수신한 Event 목록이 존재할 경우 전송하여 Event 유실을 예방
-
-//        if (hasLostData(lastEventId)) {
-//            sendLostData(lastEventId, memberId, emitterId, emitter);
-//        }
-
-        return emitter;
-    }
-
-    private String makeTimeIncludeId(Long memberId) {
-        return memberId + "_" + System.currentTimeMillis();
-    }
-
-
-
-    private void sendNotification(SseEmitter emitter, String eventId, String emitterId, Object data) {
-        try {
-            emitter.send(SseEmitter.event()
-                    .id(eventId)
-                    .data(data));
-        } catch (IOException exception) {
-            emitterRepository.deleteById(emitterId);
-        }
-    }
-
-    private boolean hasLostData(String lastEventId) {
-        return !lastEventId.isEmpty();
-    }
-
-    private void sendLostData(String lastEventId, Long memberId, String emitterId, SseEmitter emitter) {
-        Map<String, Object> eventCaches = emitterRepository.findAllEventCacheStartWithByMemberId(String.valueOf(memberId));
-        eventCaches.entrySet().stream()
-                .filter(entry -> lastEventId.compareTo(entry.getKey()) < 0)
-                .forEach(entry -> sendNotification(emitter, entry.getKey(), emitterId, entry.getValue()));
-    }
-
-    public void send(Account receiver, AlarmType notificationType, String content, String data) {
-        Notification notification = notificationRepository.save(createNotification(receiver, notificationType, content, data));
-
-        String receiverId = String.valueOf(receiver.getId());
-        String eventId = receiverId + "_" + System.currentTimeMillis();
-        Map<String, SseEmitter> emitters = emitterRepository.findAllEmitterStartWithByMemberId(receiverId);
-        emitters.forEach(
-                (key, emitter) -> {
-                    emitterRepository.saveEventCache(key, notification);
-                    sendNotification(emitter, eventId, key, NotificationResponseDto.create(notification));
-                }
-        );
-    }
-
-    private Notification createNotification(Account receiver, AlarmType notificationType, String content, String data) {
-        return Notification.builder()
-                .receiver(receiver)
-                .alarmType(notificationType)
-                .message(content)
-                .data(data)
-                .readState(false)
-                .build();
-    }
-
-
-}
